refactor(daemon): extract SSE/AMR error payloads + project-display-status
Strangler-fig slice 3: two small byte-identical preamble extractions.

- http/error-payloads.ts: createSseErrorPayload, rewriteKnownAgentStreamError,
  createAmrModelUnavailablePayload (pure; built on the typed createCompatApiError
  from http/api-errors). server.ts imports all three back for the chat-run deps.
- project-display-status.ts: normalizeProjectDisplayStatus,
  composeProjectDisplayStatus (pure). server.ts imports both back and re-exports
  them to preserve its public surface.

No behavior change. Orphan/duplicate-binding checks clean.
