+++ apps/daemon/src/http/error-payloads.ts
@@ -1,0 +1,58 @@
+// Authors: Leon Aburime using Claude Fable 5
+// @ts-nocheck — carried over verbatim from server.ts's file-level @ts-nocheck.
+// The moved bodies are untyped JS-in-TS; typing them is a later effort and new
+// sibling code must NOT copy this.
+/** @module http/error-payloads
+ * SSE/agent error payload builders.
+ *
+ * `createSseErrorPayload` wraps a code+message into the `{ message, error }`
+ * shape the chat-run SSE stream emits (built on the typed `createCompatApiError`
+ * from ./api-errors); `rewriteKnownAgentStreamError` maps a known opencode
+ * "bufio.scanner: token too long" upstream failure to a friendly retry message;
+ * `createAmrModelUnavailablePayload` builds the AMR-model-unavailable SSE error.
+ * All are pure — they read only their arguments and the imported factory.
+ *
+ * Extracted verbatim from apps/daemon/src/server.ts (strangler-fig slice 3).
+ * server.ts imports all three back for the chat-run deps object.
+ */
+
+import { createCompatApiError } from './api-errors.js';
+
+export function createSseErrorPayload(code, message, init = {}) {
+  return { message, error: createCompatApiError(code, message, init) };
+}
+
+export function rewriteKnownAgentStreamError(agentId, message, failureText = '') {
+  const rawMessage =
+    typeof message === 'string' && message.trim()
+      ? message.trim()
+      : 'Agent stream error';
+  const combined = `${rawMessage}\n${failureText}`;
+  if (
+    /bufio\.scanner:\s*token too long/i.test(combined) &&
+    /opencode/i.test(combined) &&
+    (agentId === 'opencode' || agentId === 'mimo' || agentId === 'amr' || /json-rpc id \d+/i.test(combined))
+  ) {
+    return 'The run failed due to an unknown upstream streaming error. Please retry.';
+  }
+  return rawMessage;
+}
+
+export function createAmrModelUnavailablePayload(model, init = {}) {
+  const modelText = typeof model === 'string' && model.trim()
+    ? `"${model.trim()}"`
+    : 'the selected model';
+  return createSseErrorPayload(
+    'AMR_MODEL_UNAVAILABLE',
+    `AMR model ${modelText} is not available from Vela. Refresh the AMR model list, choose a supported model, and retry this run.`,
+    {
+      retryable: false,
+      details: {
+        kind: 'amr_model',
+        action: 'choose_model',
+        ...(typeof model === 'string' && model.trim() ? { model: model.trim() } : {}),
+        ...init,
+      },
+    },
+  );
+}
